Fixed improper YogaUsage with YGNode and added opengl cleanup functions.

+++ include/Utils/Color.h
@@ -6,9 +6,9 @@ namespace bGUI {
 	class ColorUtils {
 	public:
 		// assumes hsl values are between 0.0 and 1.0
-		static Vector3f convertHSLtoRGB(float h, float s, float l);
+		static Vector3i convertHSLtoRGB(float h, float s, float l);
 		// takes in a hexadecimal number expressed as a string and converts it to rgb.
-		static Vector3f convertHEXtoRGB(const char* hexcode);
-		static Vector3f convertHEXtoRGB(const int hexcode);
+		static Vector3i convertHEXtoRGB(const char* hexcode);
+		static Vector3i convertHEXtoRGB(const int hexcode);
 	};
 }
